% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_matrix.R
\name{association.matrix}
\alias{association.matrix}
\title{Create association matrix}
\usage{
association.matrix(dyads, var)
}
\arguments{
\item{dyads}{desc}

\item{var}{desc}
}
\value{
desc
}
\description{
Create association matrix
}
