% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_index.R
\name{association_index}
\alias{association_index}
\title{Compute association indices}
\usage{
association_index(dyads)
}
\arguments{
\item{dyads}{desc}
}
\value{
desc
}
\description{
Compute association indices
}
