% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_history.R
\name{capture_history}
\alias{capture_history}
\title{Capture history}
\usage{
capture_history(ILV, groups, samp.period = "year")
}
\arguments{
\item{ILV}{desc}

\item{groups}{desc}

\item{samp.period}{desc}
}
\value{
desc
}
\description{
Capture history
}
