% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_ILV.R
\name{compile_ILV}
\alias{compile_ILV}
\title{Compile Individual Level Variables}
\usage{
compile_ILV(events)
}
\arguments{
\item{events}{desc}
}
\value{
desc
}
\description{
Compile Individual Level Variables
}
