% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_quality_score.R
\name{find_quality_score}
\alias{find_quality_score}
\title{Find quality score}
\usage{
find_quality_score(log, epath)
}
\arguments{
\item{log}{desc}

\item{epath}{desc}
}
\value{
desc
}
\description{
Find quality score
}
