% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Setup your catRlog folder system}
\usage{
setup_project()
}
\value{
Several folders are created within your photo-ID project directory.
}
\description{
Do this when you are setting up a new project. Run this function with your working directory set to your photo-ID project folder.
}
